fix go
